poker/requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
fastapi==0.115.6
fastcore==1.7.26
fastlite==0.0.13
ffmpy==0.4.0
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==5.9.0
gradio_client==1.5.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonalias==0.1.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.18.3
numpy==2.2.0
oauthlib==3.2.2
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
protobuf==5.29.1
pyarrow==18.1.0
pycoingecko==3.2.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.10.1
python-multipart==0.0.19
python-telegram-bot==21.9
pytz==2024.2
pywebio==1.8.3
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.3
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
solana==0.36.1
solders==0.23.0
soupsieve==2.6
sqlite-minutils==4.0.3
starlette==0.41.3
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
treys==0.1.8
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.2.3
user-agents==2.2.0
uvicorn==0.32.1
watchdog==6.0.0
watchfiles==1.0.3
websockets==12.0
Werkzeug==3.1.3
